+++ ios/Classes/AudioBackgroundServicePlugin.h
@@ -1,0 +1,16 @@
+//
+//  AudioBackgroundServicePlugin.h
+//  audio_service
+//
+//  Created by Chuong Vu Duy on 7/30/19.
+//
+
+#import <Flutter/Flutter.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface AudioBackgroundServicePlugin : NSObject<FlutterPlugin>
+
+@end
+
+NS_ASSUME_NONNULL_END