+++ ios/Classes/Constants.h
@@ -1,0 +1,20 @@
+//
+//  Constants.h
+//  Runner
+//
+//  Created by Chuong Vu Duy on 10/5/19.
+//  Copyright © 2019 The Chromium Authors. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+extern NSString *_Nonnull const AUDIO_CHANGED_MEDIA;
+extern NSString *_Nonnull const MEDIA_ACTION_CHANGE_STATE;
+extern NSString *_Nonnull const MEDIA_ACTION_PLAY;
+extern NSString *_Nonnull const MEDIA_ACTION_PAUSE;
+extern NSString *_Nonnull const MEDIA_ACTION_STOP;
+extern NSString *_Nonnull const MEDIA_ACTION_CHANGE;
+
+NS_ASSUME_NONNULL_END