+++ ios/Classes/Constants.m
@@ -1,0 +1,16 @@
+//
+//  Constants.m
+//  Runner
+//
+//  Created by Chuong Vu Duy on 10/5/19.
+//  Copyright © 2019 The Chromium Authors. All rights reserved.
+//
+
+#import "Constants.h"
+
+NSString *_Nonnull const AUDIO_CHANGED_MEDIA = @"audio_changed_media";
+NSString *_Nonnull const MEDIA_ACTION_CHANGE_STATE = @"media_action_change_state";
+NSString *_Nonnull const MEDIA_ACTION_PLAY = @"media_action_play";
+NSString *_Nonnull const MEDIA_ACTION_PAUSE = @"media_action_pause";
+NSString *_Nonnull const MEDIA_ACTION_STOP = @"media_action_stop";
+NSString *_Nonnull const MEDIA_ACTION_CHANGE = @"media_action_change";