+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.4.1
+
+* Fix queue support bug
+
+## 0.4.0
+
+* Breaking change: AudioServiceBackground.run takes a single parameter.
+
+## 0.3.1
+
+* Update example to disconnect when pressing back button.
+
 ## 0.3.0
 
 * Breaking change: updateTime now measured since epoch instead of boot time.